add chat component
